+++ recipes/default.rb
@@ -10,6 +10,7 @@
 # Copy some parameters into explicit Phabricator configuration
 node.default['phabricator']['config']['repository.default-local-path'] = node['phabricator']['repository_path']
 node.default['phabricator']['config']['phd.user'] = node['phabricator']['user']
+node.default['phabricator']['config']['storage.local-disk.path'] = node['phabricator']['file_storage_path']
 
 # Some default Phabricator settings
 node.default['phabricator']['config']['metamta.domain'] = node['phabricator']['domain']
@@ -119,6 +120,14 @@ end
 end
 
 directory node['phabricator']['repository_path'] do
+    action :create
+    user node['phabricator']['user']
+    group node['phabricator']['group']
+    mode "0750"
+end
+
+# Set up file storage
+directory node['phabricator']['file_storage_path'] do
     action :create
     user node['phabricator']['user']
     group node['phabricator']['group']